feat: add adaptive
