+++ client/src/App.css
@@ -10,7 +10,7 @@
 
 
 h1 {
-  font-size: 1.5rem;
+  font-size: 1rem;
 }
 .App {
   display: flex;
@@ -28,7 +28,6 @@ h1 {
   display: flex;
   justify-content: center;
   align-items: center;
-  padding: 1rem;
   height: calc(100% - (var(--headerHeight) * 2));
   min-width: 750px;
   background-image: url('./assets/img/background.jpg');
@@ -43,7 +42,7 @@ h1 {
   display: flex;
   justify-content: center;
   align-items: center;
-  padding: 1rem;
+ 
 }
 
 .main-content form {
@@ -92,9 +91,9 @@ h1 {
 .game-field {
   display: grid;
   grid-template-columns: repeat(6, 1fr);
-  grid-template-rows: repeat(3, 1fr);
+ 
   column-gap: 20px;
-  row-gap: 20px;
+  row-gap: 10px;
   
 }
 
@@ -180,3 +179,34 @@ h1 {
   background-color: #fff;
 }
 
+
+@media (max-width: 1023px) {
+
+  .main-content {
+    max-width: 600px;
+    min-width: none;
+  }
+}
+
+@media (max-width: 767px) {
+  .main-content {
+    max-width: 300px;
+    height: auto;
+    display: flex;
+    min-width: 85%;
+    padding: 1rem;
+  }
+
+  .main-content .wrapper {
+    display: flex;
+    flex-direction: column;
+  }
+  .main-content .game-field {
+    grid-template-columns: repeat(5, 1fr);
+    padding: 2rem;
+  }
+
+  .main-content .card-container {
+    max-width: 80px;
+  }
+}
